Moving test_upload_cancel out
